requirements.txt
playwright
esprima
requests
PyGithub
rich
click
flask
fastapi
uvicorn
python-dotenv
beautifulsoup4
pytest
lxml
